# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SipWiki Drinking Games API",
    "description": "Search and discover 100+ drinking games with complete rules. Find party games by players, materials, and intensity level. Includes classics like Kings Cup, Beer Pong, Flip Cup, Never Have I Ever, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "SipWiki Support",
      "email": "support@sipwiki.com",
      "url": "https://sipwiki.com"
    }
  },
  "servers": [
    {
      "url": "https://sipwiki.com/api"
    }
  ],
  "paths": {
    "/mcp": {
      "post": {
        "operationId": "mcpRequest",
        "summary": "Model Context Protocol endpoint for ChatGPT",
        "description": "Handles MCP JSON-RPC requests for drinking game tools",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": { "type": "string" },
                  "method": { "type": "string" },
                  "params": { "type": "object" },
                  "id": { "type": ["string", "number"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getServerInfo",
        "summary": "Get SipWiki MCP server information",
        "description": "Returns available tools and total game count",
        "responses": {
          "200": {
            "description": "Server info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "version": { "type": "string" },
                    "tools": { "type": "array" },
                    "total_games": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Game name (e.g., Kings Cup, Beer Pong)" },
          "slug": { "type": "string", "description": "URL-friendly identifier" },
          "description": { "type": "string", "description": "Short game description" },
          "rules": { "type": "string", "description": "Complete game rules and instructions" },
          "players": { "type": "string", "description": "Player count range (e.g., 4-10)" },
          "materials": { "type": "array", "items": { "type": "string" }, "description": "Required materials (cards, dice, cups, etc.)" },
          "sipFactor": { "type": "integer", "minimum": 1, "maximum": 5, "description": "Drinking intensity 1-5" },
          "alcoholType": { "type": "string", "enum": ["beer", "liquor", "any"] }
        }
      }
    }
  }
}